+++ src/components/common.ts
@@ -16,6 +16,9 @@ export type RefCallback<Element extends HTMLElement> = (
 
 export type Omit<T, K> = Pick<T, Exclude<keyof T, K>>;
 
+/**
+ * Wraps Object.keys with proper typescript definition of the resulting array
+ */
 export function keysOf<T, K extends keyof T>(obj: T): K[] {
   return Object.keys(obj) as K[];
 }