+++ src/index.d.ts
@@ -1,3 +1,2 @@
 /// <reference path="./components/index.d.ts" />
 /// <reference path="./themes/index.d.ts" />
-/// <reference path="./test/index.d.ts" />