+++ src/test/index.ts
@@ -1,0 +1,7 @@
+export { requiredProps } from './required_props';
+export { takeMountedSnapshot } from './take_mounted_snapshot';
+export { findTestSubject } from './find_test_subject';
+export {
+  startThrowingReactWarnings,
+  stopThrowingReactWarnings,
+} from './react_warnings';